Update hf-bitsandbytes-integration.md
Great work, as always, Younes!

I really like the additions. It makes the procedure and how you integrated it into the HF libraries very tangible. 

One major change I think needs to be made is the example about how to use a PyTorch module with bitsandbytes in Int8. One thing is, that one can train in any precision and also store the checkpoint in any precision. FP16 is not required. What is required is to use FP16 inputs into the module once it is converted with bitsandbytes. This is so, because bnb casts any model to FP16, but it will not cast the inputs to FP16 (the user has to do this). 

Otherwise, I made some small edits here and there. I will add the link to the paper once it is announced in the evening.